CoursWork_BD 1.2

+++ src/main/java/pojo/Role.java
@@ -9,7 +9,7 @@ import java.util.List;
 public class Role {
     @Id()
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long roleId;
+    private Integer roleId;
     @Column(name = "role")
     private String role;
     @Column(name = "dt_update")
